Relocate amazon IAM code
